Remove KnownNetworkLibDeps and IsKnownNetworkLibrary
KnownNetworkLibDeps (SymbolAnalysisData) was a precomputed list of
known-network-library SONames stored in the executable's record.
The runner used it to classify binaries as network-capable when they
linked against libssl, libcurl, language runtimes, etc.

This field is now redundant: the dynlib analysis pipeline
(checkDynLibDepsNetwork) already covers both cases:
- Protocol libs (libssl, libcurl): DetectedSymbols in dynlib analysis
- Language runtimes (libruby, libpython): DynamicLoadSymbols via dlopen

Changes:
- Delete known_network_libs.go and known_network_libs_test.go
- Move matchesKnownPrefix (shared helper) to syscall_wrapper_libs.go
- Update machoanalyzer test comment that referenced KnownNetworkLibDeps
